requirements.txt
discord.py>=2.5.1
Flask>=2.2.0
requests>=2.28.0
replit>=3.2.0
